Phase 1: extract chat bubbles and tool renderers from SidebarChat
Move message components, plan/review UI, builtin tool renderers, and tool headers into dedicated chat/ and tools/ modules so SidebarChat becomes a thin composer shell.
