+++ src/vs/workbench/contrib/cortexide/browser/react/src/sidebar-tsx/SidebarChat.tsx
@@ -50,6 +50,8 @@ import { PDFMessageRenderer } from '../util/PDFMessageRenderer.js';
 import { IconX, IconWarning, IconLoading, TypingCursor } from './shared/icons.js';
 import { getBasename, getFolderName, getRelative, voidOpenFileFn } from './shared/pathUtils.js';
 import { ToolChildrenWrapper, CodeChildren, ListableToolItem } from './tools/ToolPrimitives.js';
+import { ChatBubble } from './chat/ChatBubble.js';
+import { EditToolSoFar } from './tools/ToolRenderers.js';
 
 // Re-export shared modules for existing consumers (will migrate imports over time).
 export { IconX, IconWarning, IconLoading, TypingCursor } from './shared/icons.js';
@@ -57,2631 +59,11 @@ export { getBasename, getFolderName, getRelative, voidOpenFileFn } from './share
 export { ToolChildrenWrapper, CodeChildren, ListableToolItem } from './tools/ToolPrimitives.js';
 export { VoidChatArea, ButtonSubmit, ButtonStop } from './composer/VoidChatArea.js';
 export { SelectedFiles } from './composer/SelectedFiles.js';
+export { ChatBubble } from './chat/ChatBubble.js';
 
 
 
-type ToolHeaderParams = {
-	icon?: React.ReactNode;
-	title: React.ReactNode;
-	desc1: React.ReactNode;
-	desc1OnClick?: () => void;
-	desc2?: React.ReactNode;
-	isError?: boolean;
-	info?: string;
-	desc1Info?: string;
-	isRejected?: boolean;
-	numResults?: number;
-	hasNextPage?: boolean;
-	children?: React.ReactNode;
-	bottomChildren?: React.ReactNode;
-	onClick?: () => void;
-	desc2OnClick?: () => void;
-	isOpen?: boolean;
-	className?: string;
-}
 
-const ToolHeaderWrapper = ({
-	icon,
-	title,
-	desc1,
-	desc1OnClick,
-	desc1Info,
-	desc2,
-	numResults,
-	hasNextPage,
-	children,
-	info,
-	bottomChildren,
-	isError,
-	onClick,
-	desc2OnClick,
-	isOpen,
-	isRejected,
-	className, // applies to the main content
-}: ToolHeaderParams) => {
-
-	const [isOpen_, setIsOpen] = useState(false);
-	const isExpanded = isOpen !== undefined ? isOpen : isOpen_
-
-	const isDropdown = children !== undefined // null ALLOWS dropdown
-	const isClickable = !!(isDropdown || onClick)
-
-	const isDesc1Clickable = !!desc1OnClick
-
-	const desc1HTML = <span
-		className={`text-void-fg-4 text-xs italic truncate ml-2
-			${isDesc1Clickable ? 'cursor-pointer hover:brightness-125 transition-all duration-150' : ''}
-		`}
-		onClick={desc1OnClick}
-		{...desc1Info ? {
-			'data-tooltip-id': 'cortex-tooltip',
-			'data-tooltip-content': desc1Info,
-			'data-tooltip-place': 'top',
-			'data-tooltip-delay-show': 1000,
-		} : {}}
-	>{desc1}</span>
-
-	return (<div className=''>
-		<div className={`w-full border border-void-border-3 rounded px-2 py-1 bg-void-bg-3 overflow-hidden ${className}`}>
-			{/* header */}
-			<div className={`select-none flex items-center min-h-[24px]`}>
-				<div className={`flex items-center w-full gap-x-2 overflow-hidden justify-between ${isRejected ? 'line-through' : ''}`}>
-					{/* left */}
-					<div // container for if desc1 is clickable
-						className='ml-1 flex items-center overflow-hidden'
-					>
-						{/* title eg "> Edited File" */}
-						<div className={`
-							flex items-center min-w-0 overflow-hidden grow
-							${isClickable ? 'cursor-pointer hover:brightness-125 transition-all duration-150' : ''}
-						`}
-							onClick={() => {
-								if (isDropdown) { setIsOpen(v => !v); }
-								if (onClick) { onClick(); }
-							}}
-						>
-							{isDropdown && (<ChevronRight
-								className={`
-								text-void-fg-3 mr-0.5 h-4 w-4 flex-shrink-0 transition-transform duration-100 ease-[cubic-bezier(0.4,0,0.2,1)]
-								${isExpanded ? 'rotate-90' : ''}
-							`}
-							/>)}
-							<span className="text-void-fg-3 flex-shrink-0">{title}</span>
-
-							{!isDesc1Clickable && desc1HTML}
-						</div>
-						{isDesc1Clickable && desc1HTML}
-					</div>
-
-					{/* right */}
-					<div className="flex items-center gap-x-2 flex-shrink-0">
-
-						{info && <CircleEllipsis
-							className='ml-2 text-void-fg-4 opacity-60 flex-shrink-0'
-							size={14}
-							data-tooltip-id='cortex-tooltip'
-							data-tooltip-content={info}
-							data-tooltip-place='top-end'
-						/>}
-
-						{isError && <AlertTriangle
-							className='text-void-warning opacity-90 flex-shrink-0'
-							size={14}
-							data-tooltip-id='cortex-tooltip'
-							data-tooltip-content={'Error running tool'}
-							data-tooltip-place='top'
-						/>}
-						{isRejected && <Ban
-							className='text-void-fg-4 opacity-90 flex-shrink-0'
-							size={14}
-							data-tooltip-id='cortex-tooltip'
-							data-tooltip-content={'Canceled'}
-							data-tooltip-place='top'
-						/>}
-						{desc2 && <span className="text-void-fg-4 text-xs" onClick={desc2OnClick}>
-							{desc2}
-						</span>}
-						{numResults !== undefined && (
-							<span className="text-void-fg-4 text-xs ml-auto mr-1">
-								{`${numResults}${hasNextPage ? '+' : ''} result${numResults !== 1 ? 's' : ''}`}
-							</span>
-						)}
-					</div>
-				</div>
-			</div>
-			{/* children */}
-			{<div
-				className={`overflow-hidden transition-all duration-200 ease-in-out ${isExpanded ? 'opacity-100 py-1' : 'max-h-0 opacity-0'}
-					text-void-fg-4 rounded-sm overflow-x-auto
-				  `}
-			//    bg-black bg-opacity-10 border border-void-border-4 border-opacity-50
-			>
-				{children}
-			</div>}
-		</div>
-		{bottomChildren}
-	</div>);
-};
-
-
-
-const EditTool = ({ toolMessage, threadId, messageIdx, content }: Parameters<ResultWrapper<'edit_file' | 'rewrite_file'>>[0] & { content: string }) => {
-	const accessor = useAccessor()
-	const isError = false
-	const isRejected = toolMessage.type === 'rejected'
-
-	const title = getTitle(toolMessage)
-
-	const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-	const icon = null
-
-	const { rawParams, params, name } = toolMessage
-	const desc1OnClick = () => voidOpenFileFn(params.uri, accessor)
-	const componentParams: ToolHeaderParams = { title, desc1, desc1OnClick, desc1Info, isError, icon, isRejected, }
-
-
-	const editToolType = toolMessage.name === 'edit_file' ? 'diff' : 'rewrite'
-	if (toolMessage.type === 'running_now' || toolMessage.type === 'tool_request') {
-		componentParams.children = <ToolChildrenWrapper className='bg-void-bg-3'>
-			<EditToolChildren
-				uri={params.uri}
-				code={content}
-				type={editToolType}
-			/>
-		</ToolChildrenWrapper>
-		// JumpToFileButton removed in favor of FileLinkText
-	}
-	else if (toolMessage.type === 'success' || toolMessage.type === 'rejected' || toolMessage.type === 'tool_error') {
-		// add apply box
-		const applyBoxId = getApplyBoxId({
-			threadId: threadId,
-			messageIdx: messageIdx,
-			tokenIdx: 'N/A',
-		})
-		componentParams.desc2 = <EditToolHeaderButtons
-			applyBoxId={applyBoxId}
-			uri={params.uri}
-			codeStr={content}
-			toolName={name}
-			threadId={threadId}
-		/>
-
-		// add children
-		componentParams.children = <ToolChildrenWrapper className='bg-void-bg-3'>
-			<EditToolChildren
-				uri={params.uri}
-				code={content}
-				type={editToolType}
-			/>
-		</ToolChildrenWrapper>
-
-		if (toolMessage.type === 'success' || toolMessage.type === 'rejected') {
-			const { result } = toolMessage
-			componentParams.bottomChildren = <BottomChildren title='Lint errors'>
-				{result?.lintErrors?.map((error, i) => (
-					<div key={`${error.startLineNumber}-${error.endLineNumber}-${i}`} className='whitespace-nowrap'>Lines {error.startLineNumber}-{error.endLineNumber}: {error.message}</div>
-				))}
-			</BottomChildren>
-		}
-		else if (toolMessage.type === 'tool_error') {
-			// error
-			const { result } = toolMessage
-			componentParams.bottomChildren = <BottomChildren title='Error'>
-				<CodeChildren>
-					{result}
-				</CodeChildren>
-			</BottomChildren>
-		}
-	}
-
-	return <ToolHeaderWrapper {...componentParams} />
-}
-
-const SimplifiedToolHeader = ({
-	title,
-	children,
-}: {
-	title: string;
-	children?: React.ReactNode;
-}) => {
-	const [isOpen, setIsOpen] = useState(false);
-	const isDropdown = children !== undefined;
-	return (
-		<div>
-			<div className="w-full">
-				{/* header */}
-				<div
-					className={`select-none flex items-center min-h-[24px] ${isDropdown ? 'cursor-pointer' : ''}`}
-					onClick={() => {
-						if (isDropdown) { setIsOpen(v => !v); }
-					}}
-				>
-					{isDropdown && (
-						<ChevronRight
-							className={`text-void-fg-3 mr-0.5 h-4 w-4 flex-shrink-0 transition-transform duration-100 ease-[cubic-bezier(0.4,0,0.2,1)] ${isOpen ? 'rotate-90' : ''}`}
-						/>
-					)}
-					<div className="flex items-center w-full overflow-hidden">
-						<span className="text-void-fg-3">{title}</span>
-					</div>
-				</div>
-				{/* children */}
-				{<div
-					className={`overflow-hidden transition-all duration-200 ease-in-out ${isOpen ? 'opacity-100' : 'max-h-0 opacity-0'} text-void-fg-4`}
-				>
-					{children}
-				</div>}
-			</div>
-		</div>
-	);
-};
-
-
-
-
-const UserMessageComponent = ({ chatMessage, messageIdx, isCheckpointGhost, currCheckpointIdx, _scrollToBottom }: { chatMessage: ChatMessage & { role: 'user' }, messageIdx: number, currCheckpointIdx: number | undefined, isCheckpointGhost: boolean, _scrollToBottom: (() => void) | null }) => {
-
-	const accessor = useAccessor()
-	const chatThreadsService = accessor.get('IChatThreadService')
-
-	// Subscribe to thread state changes properly
-	const chatThreadsState = useChatThreadsState()
-	const currentThreadId = chatThreadsState.currentThreadId
-
-	// global state
-	let isBeingEdited = false
-	let stagingSelections: StagingSelectionItem[] = []
-	let setIsBeingEdited = (_: boolean) => { }
-	let setStagingSelections = (_: StagingSelectionItem[]) => { }
-
-	if (messageIdx !== undefined) {
-		const _state = chatThreadsService.getCurrentMessageState(messageIdx)
-		isBeingEdited = _state.isBeingEdited
-		stagingSelections = _state.stagingSelections
-		setIsBeingEdited = (v) => chatThreadsService.setCurrentMessageState(messageIdx, { isBeingEdited: v })
-		setStagingSelections = (s) => chatThreadsService.setCurrentMessageState(messageIdx, { stagingSelections: s })
-	}
-
-
-	// local state
-	const mode: ChatBubbleMode = isBeingEdited ? 'edit' : 'display'
-	const [isFocused, setIsFocused] = useState(false)
-	const [isHovered, setIsHovered] = useState(false)
-	const [isDisabled, setIsDisabled] = useState(false)
-	const [textAreaRefState, setTextAreaRef] = useState<HTMLTextAreaElement | null>(null)
-	const textAreaFnsRef = useRef<TextAreaFns | null>(null)
-	// initialize on first render, and when edit was just enabled
-	const _mustInitialize = useRef(true)
-	const _justEnabledEdit = useRef(false)
-	useEffect(() => {
-		const canInitialize = mode === 'edit' && textAreaRefState
-		const shouldInitialize = _justEnabledEdit.current || _mustInitialize.current
-		if (canInitialize && shouldInitialize) {
-			setStagingSelections(
-				(chatMessage.selections || []).map(s => { // quick hack so we dont have to do anything more
-					if (s.type === 'File') return { ...s, state: { ...s.state, wasAddedAsCurrentFile: false, } }
-					else return s
-				})
-			)
-
-			if (textAreaFnsRef.current)
-				textAreaFnsRef.current.setValue(chatMessage.displayContent || '')
-
-			textAreaRefState.focus();
-
-			_justEnabledEdit.current = false
-			_mustInitialize.current = false
-		}
-
-	}, [chatMessage, mode, textAreaRefState, setStagingSelections])
-
-	const onOpenEdit = () => {
-		setIsBeingEdited(true)
-		chatThreadsService.setCurrentlyFocusedMessageIdx(messageIdx)
-		_justEnabledEdit.current = true
-	}
-	const onCloseEdit = () => {
-		setIsFocused(false)
-		setIsHovered(false)
-		setIsBeingEdited(false)
-		chatThreadsService.setCurrentlyFocusedMessageIdx(undefined)
-
-	}
-
-	const EditSymbol = mode === 'display' ? Pencil : X
-
-
-	let chatbubbleContents: React.ReactNode
-	if (mode === 'display') {
-		const hasImages = chatMessage.images && chatMessage.images.length > 0;
-		const hasPDFs = chatMessage.pdfs && chatMessage.pdfs.length > 0;
-		const hasAttachments = hasImages || hasPDFs;
-
-		chatbubbleContents = <>
-			<SelectedFiles type='past' messageIdx={messageIdx} selections={chatMessage.selections || []} />
-			{hasImages && (
-				<div className="px-0.5 py-2">
-					<ImageMessageRenderer
-						images={chatMessage.images}
-					/>
-				</div>
-			)}
-			{hasPDFs && (
-				<div className="px-0.5 py-2">
-					<PDFMessageRenderer
-						pdfs={chatMessage.pdfs}
-					/>
-				</div>
-			)}
-			{chatMessage.displayContent && (
-				<span className='px-0.5'>{chatMessage.displayContent}</span>
-			)}
-		</>
-	}
-	else if (mode === 'edit') {
-
-		const onSubmit = async () => {
-
-			if (isDisabled) return;
-			if (!textAreaRefState) return;
-			if (messageIdx === undefined) return;
-
-			// cancel any streams on this thread - use subscribed state
-			const threadId = currentThreadId
-
-			// Defensive check: verify the message is still a user message before editing
-			const thread = chatThreadsState.allThreads[threadId]
-			if (!thread || !thread.messages || thread.messages[messageIdx]?.role !== 'user') {
-				console.error('Error while editing message: Message is not a user message or no longer exists')
-				setIsBeingEdited(false)
-				chatThreadsService.setCurrentlyFocusedMessageIdx(undefined)
-				return
-			}
-
-			await chatThreadsService.abortRunning(threadId)
-
-			// update state
-			setIsBeingEdited(false)
-			chatThreadsService.setCurrentlyFocusedMessageIdx(undefined)
-
-			// stream the edit
-			const userMessage = textAreaRefState.value;
-			try {
-				await chatThreadsService.editUserMessageAndStreamResponse({ userMessage, messageIdx, threadId })
-			} catch (e) {
-				console.error('Error while editing message:', e)
-			}
-			await chatThreadsService.focusCurrentChat()
-			requestAnimationFrame(() => _scrollToBottom?.())
-		}
-
-		const onAbort = async () => {
-			// use subscribed state
-			const threadId = currentThreadId
-			await chatThreadsService.abortRunning(threadId)
-		}
-
-		const onKeyDown = (e: KeyboardEvent<HTMLTextAreaElement>) => {
-			if (e.key === 'Escape') {
-				onCloseEdit()
-			}
-			if (e.key === 'Enter' && !e.shiftKey && !e.nativeEvent.isComposing) {
-				onSubmit()
-			}
-		}
-
-		if (!chatMessage.content) { // don't show if empty and not loading (if loading, want to show).
-			return null
-		}
-
-		chatbubbleContents = <VoidChatArea
-			featureName='Chat'
-			onSubmit={onSubmit}
-			onAbort={onAbort}
-			isStreaming={false}
-			isDisabled={isDisabled}
-			showSelections={true}
-			showProspectiveSelections={false}
-			selections={stagingSelections}
-			setSelections={setStagingSelections}
-		>
-			<VoidInputBox2
-				enableAtToMention
-				appearance="chatDark"
-				ref={setTextAreaRef}
-				className='min-h-[60px] px-3 py-3 rounded-2xl'
-				placeholder="Plan, @ for context, / for commands"
-				onChangeText={(text) => setIsDisabled(!text)}
-				onFocus={() => {
-					setIsFocused(true)
-					chatThreadsService.setCurrentlyFocusedMessageIdx(messageIdx);
-				}}
-				onBlur={() => {
-					setIsFocused(false)
-				}}
-				onKeyDown={onKeyDown}
-				fnsRef={textAreaFnsRef}
-				multiline={true}
-			/>
-		</VoidChatArea>
-	}
-
-	const isMsgAfterCheckpoint = currCheckpointIdx !== undefined && currCheckpointIdx === messageIdx - 1
-
-	return <div
-		// align chatbubble accoridng to role
-		className={`
-        relative ml-auto
-        ${mode === 'edit' ? 'w-full max-w-full'
-				: mode === 'display' ? `self-end w-fit max-w-full whitespace-pre-wrap` : '' // user words should be pre
-			}
-
-        ${isCheckpointGhost && !isMsgAfterCheckpoint ? 'opacity-50 pointer-events-none' : ''}
-    `}
-		onMouseEnter={() => setIsHovered(true)}
-		onMouseLeave={() => setIsHovered(false)}
-	>
-		<div
-			// style chatbubble according to role
-			className={`
-            text-left rounded-lg max-w-full
-            ${mode === 'edit' ? ''
-					: mode === 'display' ? 'p-2 flex flex-col bg-void-bg-1 text-void-fg-1 overflow-x-auto cursor-pointer' : ''
-				}
-        `}
-			onClick={() => { if (mode === 'display') { onOpenEdit() } }}
-		>
-			{chatbubbleContents}
-		</div>
-
-
-
-		<div
-			className="absolute -top-1 -right-1 translate-x-0 -translate-y-0 z-1"
-		// data-tooltip-id='cortex-tooltip'
-		// data-tooltip-content='Edit message'
-		// data-tooltip-place='left'
-		>
-			<EditSymbol
-				size={18}
-				className={`
-                    cursor-pointer
-                    p-[2px]
-                    bg-void-bg-1 border border-void-border-1 rounded-md
-                    transition-opacity duration-200 ease-in-out
-                    ${isHovered || (isFocused && mode === 'edit') ? 'opacity-100' : 'opacity-0'}
-                `}
-				onClick={() => {
-					if (mode === 'display') {
-						onOpenEdit()
-					} else if (mode === 'edit') {
-						onCloseEdit()
-					}
-				}}
-			/>
-		</div>
-
-
-	</div>
-
-}
-
-const SmallProseWrapper = ({ children }: { children: React.ReactNode }) => {
-	return <div className='
-text-void-fg-4
-prose
-prose-sm
-break-words
-max-w-none
-leading-snug
-text-[13px]
-
-[&>:first-child]:!mt-0
-[&>:last-child]:!mb-0
-
-prose-h1:text-[14px]
-prose-h1:my-4
-
-prose-h2:text-[13px]
-prose-h2:my-4
-
-prose-h3:text-[13px]
-prose-h3:my-3
-
-prose-h4:text-[13px]
-prose-h4:my-2
-
-prose-p:my-2
-prose-p:leading-snug
-prose-hr:my-2
-
-prose-ul:my-2
-prose-ul:pl-4
-prose-ul:list-outside
-prose-ul:list-disc
-prose-ul:leading-snug
-
-
-prose-ol:my-2
-prose-ol:pl-4
-prose-ol:list-outside
-prose-ol:list-decimal
-prose-ol:leading-snug
-
-marker:text-inherit
-
-prose-blockquote:pl-2
-prose-blockquote:my-2
-
-prose-code:text-void-fg-3
-prose-code:text-[12px]
-prose-code:before:content-none
-prose-code:after:content-none
-
-prose-pre:text-[12px]
-prose-pre:p-2
-prose-pre:my-2
-
-prose-table:text-[13px]
-'>
-		{children}
-	</div>
-}
-
-const ProseWrapper = ({ children }: { children: React.ReactNode }) => {
-	return <div className='
-text-void-fg-2
-prose
-prose-sm
-break-words
-prose-p:block
-prose-hr:my-4
-prose-pre:my-2
-marker:text-inherit
-prose-ol:list-outside
-prose-ol:list-decimal
-prose-ul:list-outside
-prose-ul:list-disc
-prose-li:my-0
-prose-code:before:content-none
-prose-code:after:content-none
-prose-headings:prose-sm
-prose-headings:font-bold
-
-prose-p:leading-normal
-prose-ol:leading-normal
-prose-ul:leading-normal
-
-max-w-none
-'
-	>
-		{children}
-	</div>
-}
-const AssistantMessageComponent = React.memo(({ chatMessage, isCheckpointGhost, isCommitted, messageIdx }: { chatMessage: ChatMessage & { role: 'assistant' }, isCheckpointGhost: boolean, messageIdx: number, isCommitted: boolean }) => {
-
-	const accessor = useAccessor()
-	const chatThreadsService = accessor.get('IChatThreadService')
-
-	const reasoningStr = chatMessage.reasoning?.trim() || null
-	const hasReasoning = !!reasoningStr
-	const isDoneReasoning = !!chatMessage.displayContent
-	const thread = chatThreadsService.getCurrentThread()
-
-
-	const chatMessageLocation: ChatMessageLocation = useMemo(() => ({
-		threadId: thread.id,
-		messageIdx: messageIdx,
-	}), [thread.id, messageIdx])
-
-	const isEmpty = !chatMessage.displayContent && !chatMessage.reasoning
-	if (isEmpty) return null
-
-	return <>
-		{/* reasoning token */}
-		{hasReasoning &&
-			<div className={`${isCheckpointGhost ? 'opacity-50' : ''}`}>
-				<ReasoningWrapper isDoneReasoning={isDoneReasoning} isStreaming={!isCommitted}>
-					<SmallProseWrapper>
-						<ChatMarkdownRender
-							string={reasoningStr}
-							chatMessageLocation={chatMessageLocation}
-							isApplyEnabled={false}
-							isLinkDetectionEnabled={true}
-						/>
-					</SmallProseWrapper>
-				</ReasoningWrapper>
-			</div>
-		}
-
-		{/* assistant message */}
-		{chatMessage.displayContent &&
-			<div
-				className={`${isCheckpointGhost ? 'opacity-50' : ''} ${!isCommitted ? 'streaming-content-chunk' : ''}`}
-				role={!isCommitted ? "status" : undefined}
-				aria-live={!isCommitted ? "polite" : undefined}
-				aria-atomic={!isCommitted ? "false" : undefined}
-			>
-				<ProseWrapper>
-					<ChatMarkdownRender
-						string={chatMessage.displayContent || ''}
-						chatMessageLocation={chatMessageLocation}
-						isApplyEnabled={true}
-						isLinkDetectionEnabled={true}
-					/>
-					{!isCommitted && <TypingCursor className="text-void-fg-2" aria-label="Streaming content" />}
-				</ProseWrapper>
-			</div>
-		}
-	</>
-
-}, (prev, next) => {
-	// Custom comparison: only re-render if message content, checkpoint state, or committed state changes
-	return prev.chatMessage.displayContent === next.chatMessage.displayContent &&
-		prev.chatMessage.reasoning === next.chatMessage.reasoning &&
-		prev.isCheckpointGhost === next.isCheckpointGhost &&
-		prev.isCommitted === next.isCommitted &&
-		prev.messageIdx === next.messageIdx
-})
-
-const ReasoningWrapper = ({ isDoneReasoning, isStreaming, children }: { isDoneReasoning: boolean, isStreaming: boolean, children: React.ReactNode }) => {
-	const isDone = isDoneReasoning || !isStreaming
-	const isWriting = !isDone
-	const [isOpen, setIsOpen] = useState(isWriting)
-	useEffect(() => {
-		if (!isWriting) setIsOpen(false) // if just finished reasoning, close
-	}, [isWriting])
-	return <ToolHeaderWrapper title='Reasoning' desc1={isWriting ? <IconLoading state="thinking" inline /> : ''} isOpen={isOpen} onClick={() => setIsOpen(v => !v)}>
-		<ToolChildrenWrapper>
-			<div className='!select-text cursor-auto'>
-				{children}
-			</div>
-		</ToolChildrenWrapper>
-	</ToolHeaderWrapper>
-}
-
-
-
-
-// should either be past or "-ing" tense, not present tense. Eg. when the LLM searches for something, the user expects it to say "I searched for X" or "I am searching for X". Not "I search X".
-
-const loadingTitleWrapper = (item: React.ReactNode): React.ReactNode => {
-	return <span className='flex items-center flex-nowrap gap-1'>
-		{item}
-		<IconLoading state="processing" inline className='w-3 text-sm' />
-	</span>
-}
-
-const titleOfBuiltinToolName = {
-	'read_file': { done: 'Read file', proposed: 'Read file', running: loadingTitleWrapper('Reading file') },
-	'ls_dir': { done: 'Inspected folder', proposed: 'Inspect folder', running: loadingTitleWrapper('Inspecting folder') },
-	'get_dir_tree': { done: 'Inspected folder tree', proposed: 'Inspect folder tree', running: loadingTitleWrapper('Inspecting folder tree') },
-	'search_pathnames_only': { done: 'Searched by file name', proposed: 'Search by file name', running: loadingTitleWrapper('Searching by file name') },
-	'search_for_files': { done: 'Searched', proposed: 'Search', running: loadingTitleWrapper('Searching') },
-	'create_file_or_folder': { done: `Created`, proposed: `Create`, running: loadingTitleWrapper(`Creating`) },
-	'delete_file_or_folder': { done: `Deleted`, proposed: `Delete`, running: loadingTitleWrapper(`Deleting`) },
-	'edit_file': { done: `Edited file`, proposed: 'Edit file', running: loadingTitleWrapper('Editing file') },
-	'rewrite_file': { done: `Wrote file`, proposed: 'Write file', running: loadingTitleWrapper('Writing file') },
-	'run_command': { done: `Ran terminal`, proposed: 'Run terminal', running: loadingTitleWrapper('Running terminal') },
-	'run_persistent_command': { done: `Ran terminal`, proposed: 'Run terminal', running: loadingTitleWrapper('Running terminal') },
-
-	'open_persistent_terminal': { done: `Opened terminal`, proposed: 'Open terminal', running: loadingTitleWrapper('Opening terminal') },
-	'kill_persistent_terminal': { done: `Killed terminal`, proposed: 'Kill terminal', running: loadingTitleWrapper('Killing terminal') },
-
-	'read_lint_errors': { done: `Read lint errors`, proposed: 'Read lint errors', running: loadingTitleWrapper('Reading lint errors') },
-	'search_in_file': { done: 'Searched in file', proposed: 'Search in file', running: loadingTitleWrapper('Searching in file') },
-	'web_search': { done: 'Searched the web', proposed: 'Search the web', running: loadingTitleWrapper('Searching the web') },
-	'browse_url': { done: 'Fetched web page', proposed: 'Fetch web page', running: loadingTitleWrapper('Fetching web page') },
-} as const satisfies Record<BuiltinToolName, { done: any, proposed: any, running: any }>
-
-
-const getTitle = (toolMessage: Pick<ChatMessage & { role: 'tool' }, 'name' | 'type' | 'mcpServerName'>): React.ReactNode => {
-	const t = toolMessage
-
-	// non-built-in title
-	if (!builtinToolNames.includes(t.name as BuiltinToolName)) {
-		// descriptor of Running or Ran etc
-		const descriptor =
-			t.type === 'success' ? 'Called'
-				: t.type === 'running_now' ? 'Calling'
-					: t.type === 'tool_request' ? 'Call'
-						: t.type === 'rejected' ? 'Call'
-							: t.type === 'invalid_params' ? 'Call'
-								: t.type === 'tool_error' ? 'Call'
-									: 'Call'
-
-
-		const title = `${descriptor} ${toolMessage.mcpServerName || 'MCP'}`
-		if (t.type === 'running_now' || t.type === 'tool_request')
-			return loadingTitleWrapper(title)
-		return title
-	}
-
-	// built-in title
-	else {
-		const toolName = t.name as BuiltinToolName
-		const entry = titleOfBuiltinToolName[toolName]
-		// Defensive: a name can pass `builtinToolNames.includes()` (a runtime list)
-		// yet be missing from this title map — e.g. type/runtime drift, or a weak/free
-		// model emitting an unexpected built-in tool name. A missing entry must never
-		// crash the whole chat render (it previously threw "reading 'proposed'").
-		if (!entry) {
-			const verb = t.type === 'success' ? 'Called' : t.type === 'running_now' ? 'Calling' : 'Call'
-			const fallback = `${verb} ${toolName}`
-			return t.type === 'running_now' ? loadingTitleWrapper(fallback) : fallback
-		}
-		if (t.type === 'success') return entry.done
-		if (t.type === 'running_now') return entry.running
-		return entry.proposed
-	}
-}
-
-
-const toolNameToDesc = (toolName: BuiltinToolName, _toolParams: BuiltinToolCallParams[BuiltinToolName] | undefined, accessor: ReturnType<typeof useAccessor>): {
-	desc1: React.ReactNode,
-	desc1Info?: string,
-} => {
-
-	if (!_toolParams) {
-		return { desc1: '', };
-	}
-
-	const x = {
-		'read_file': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['read_file']
-			return {
-				desc1: getBasename(toolParams.uri.fsPath),
-				desc1Info: getRelative(toolParams.uri, accessor),
-			};
-		},
-		'ls_dir': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['ls_dir']
-			return {
-				desc1: getFolderName(toolParams.uri.fsPath),
-				desc1Info: getRelative(toolParams.uri, accessor),
-			};
-		},
-		'search_pathnames_only': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['search_pathnames_only']
-			return {
-				desc1: `"${toolParams.query}"`,
-			}
-		},
-		'search_for_files': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['search_for_files']
-			return {
-				desc1: `"${toolParams.query}"`,
-			}
-		},
-		'search_in_file': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['search_in_file'];
-			return {
-				desc1: `"${toolParams.query}"`,
-				desc1Info: getRelative(toolParams.uri, accessor),
-			};
-		},
-		'create_file_or_folder': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['create_file_or_folder']
-			return {
-				desc1: toolParams.isFolder ? getFolderName(toolParams.uri.fsPath) ?? '/' : getBasename(toolParams.uri.fsPath),
-				desc1Info: getRelative(toolParams.uri, accessor),
-			}
-		},
-		'delete_file_or_folder': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['delete_file_or_folder']
-			return {
-				desc1: toolParams.isFolder ? getFolderName(toolParams.uri.fsPath) ?? '/' : getBasename(toolParams.uri.fsPath),
-				desc1Info: getRelative(toolParams.uri, accessor),
-			}
-		},
-		'rewrite_file': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['rewrite_file']
-			return {
-				desc1: getBasename(toolParams.uri.fsPath),
-				desc1Info: getRelative(toolParams.uri, accessor),
-			}
-		},
-		'edit_file': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['edit_file']
-			return {
-				desc1: getBasename(toolParams.uri.fsPath),
-				desc1Info: getRelative(toolParams.uri, accessor),
-			}
-		},
-		'run_command': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['run_command']
-			return {
-				desc1: `"${toolParams.command}"`,
-			}
-		},
-		'run_persistent_command': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['run_persistent_command']
-			return {
-				desc1: `"${toolParams.command}"`,
-			}
-		},
-		'open_persistent_terminal': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['open_persistent_terminal']
-			return { desc1: '' }
-		},
-		'kill_persistent_terminal': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['kill_persistent_terminal']
-			return { desc1: toolParams.persistentTerminalId }
-		},
-		'get_dir_tree': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['get_dir_tree']
-			return {
-				desc1: getFolderName(toolParams.uri.fsPath) ?? '/',
-				desc1Info: getRelative(toolParams.uri, accessor),
-			}
-		},
-		'read_lint_errors': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['read_lint_errors']
-			return {
-				desc1: getBasename(toolParams.uri.fsPath),
-				desc1Info: getRelative(toolParams.uri, accessor),
-			}
-		},
-		'web_search': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['web_search']
-			return {
-				desc1: `"${toolParams.query}"`,
-			}
-		},
-		'browse_url': () => {
-			const toolParams = _toolParams as BuiltinToolCallParams['browse_url']
-			return {
-				desc1: toolParams.url,
-				desc1Info: new URL(toolParams.url).hostname,
-			}
-		}
-	}
-
-	try {
-		return x[toolName]?.() || { desc1: '' }
-	}
-	catch {
-		return { desc1: '' }
-	}
-}
-
-const ToolRequestAcceptRejectButtons = ({ toolName }: { toolName: ToolName }) => {
-	const accessor = useAccessor()
-	const chatThreadsService = accessor.get('IChatThreadService')
-	const metricsService = accessor.get('IMetricsService')
-	const cortexideSettingsService = accessor.get('ICortexideSettingsService')
-	const voidSettingsState = useSettingsState()
-
-	// Subscribe to thread state changes properly
-	const chatThreadsState = useChatThreadsState()
-	const currentThreadId = chatThreadsState.currentThreadId
-
-	const onAccept = useCallback(() => {
-		try { // this doesn't need to be wrapped in try/catch anymore
-			// use subscribed state
-			chatThreadsService.approveLatestToolRequest(currentThreadId)
-			metricsService.capture('Tool Request Accepted', {})
-		} catch (e) { console.error('Error while approving message in chat:', e) }
-	}, [chatThreadsService, metricsService, currentThreadId])
-
-	const onReject = useCallback(() => {
-		try {
-			// use subscribed state
-			chatThreadsService.rejectLatestToolRequest(currentThreadId)
-		} catch (e) { console.error('Error while approving message in chat:', e) }
-		metricsService.capture('Tool Request Rejected', {})
-	}, [chatThreadsService, metricsService, currentThreadId])
-
-	const approveButton = (
-		<button
-			onClick={onAccept}
-			className={`
-                px-2 py-1
-                bg-[var(--vscode-button-background)]
-                text-[var(--vscode-button-foreground)]
-                hover:bg-[var(--vscode-button-hoverBackground)]
-                rounded
-                text-sm font-medium
-            `}
-		>
-			Approve
-		</button>
-	)
-
-	const cancelButton = (
-		<button
-			onClick={onReject}
-			className={`
-                px-2 py-1
-                bg-[var(--vscode-button-secondaryBackground)]
-                text-[var(--vscode-button-secondaryForeground)]
-                hover:bg-[var(--vscode-button-secondaryHoverBackground)]
-                rounded
-                text-sm font-medium
-            `}
-		>
-			Cancel
-		</button>
-	)
-
-	const approvalType = isABuiltinToolName(toolName) ? approvalTypeOfBuiltinToolName[toolName] : 'MCP tools'
-	const approvalToggle = approvalType ? <div key={approvalType} className="flex items-center ml-2 gap-x-1">
-		<ToolApprovalTypeSwitch size='xs' approvalType={approvalType} desc={`Auto-approve ${approvalType}`} />
-	</div> : null
-
-	return <div className="flex gap-2 mx-0.5 items-center">
-		{approveButton}
-		{cancelButton}
-		{approvalToggle}
-	</div>
-}
-
-const EditToolChildren = ({ uri, code, type }: { uri: URI | undefined, code: string, type: 'diff' | 'rewrite' }) => {
-
-	const content = type === 'diff' ?
-		<VoidDiffEditor uri={uri} searchReplaceBlocks={code} />
-		: <ChatMarkdownRender string={`\`\`\`\n${code}\n\`\`\``} codeURI={uri} chatMessageLocation={undefined} />
-
-	return <div className='!select-text cursor-auto'>
-		<SmallProseWrapper>
-			{content}
-		</SmallProseWrapper>
-	</div>
-
-}
-
-
-const LintErrorChildren = ({ lintErrors }: { lintErrors: LintErrorItem[] }) => {
-	return <div className="text-xs text-void-fg-4 opacity-80 border-l-2 border-void-warning px-2 py-0.5 flex flex-col gap-0.5 overflow-x-auto whitespace-nowrap">
-		{lintErrors.map((error, i) => (
-			<div key={`${error.startLineNumber}-${error.endLineNumber}-${i}`}>Lines {error.startLineNumber}-{error.endLineNumber}: {error.message}</div>
-		))}
-	</div>
-}
-
-const BottomChildren = ({ children, title }: { children: React.ReactNode, title: string }) => {
-	const [isOpen, setIsOpen] = useState(false);
-	if (!children) return null;
-	return (
-		<div className="w-full px-2 mt-0.5">
-			<div
-				className={`flex items-center cursor-pointer select-none transition-colors duration-150 pl-0 py-0.5 rounded group`}
-				onClick={() => setIsOpen(o => !o)}
-				style={{ background: 'none' }}
-			>
-				<ChevronRight
-					className={`mr-1 h-3 w-3 flex-shrink-0 transition-transform duration-100 text-void-fg-4 group-hover:text-void-fg-3 ${isOpen ? 'rotate-90' : ''}`}
-				/>
-				<span className="font-medium text-void-fg-4 group-hover:text-void-fg-3 text-xs">{title}</span>
-			</div>
-			<div
-				className={`overflow-hidden transition-all duration-200 ease-in-out ${isOpen ? 'opacity-100' : 'max-h-0 opacity-0'} text-xs pl-4`}
-			>
-				<div className="overflow-x-auto text-void-fg-4 opacity-90 border-l-2 border-void-warning px-2 py-0.5">
-					{children}
-				</div>
-			</div>
-		</div>
-	);
-}
-
-
-const EditToolHeaderButtons = ({ applyBoxId, uri, codeStr, toolName, threadId }: { threadId: string, applyBoxId: string, uri: URI, codeStr: string, toolName: 'edit_file' | 'rewrite_file' }) => {
-	const { streamState } = useEditToolStreamState({ applyBoxId, uri })
-	return <div className='flex items-center gap-1'>
-		{/* <StatusIndicatorForApplyButton applyBoxId={applyBoxId} uri={uri} /> */}
-		{/* <JumpToFileButton uri={uri} /> */}
-		{streamState === 'idle-no-changes' && <CopyButton codeStr={codeStr} toolTipName='Copy' />}
-		<EditToolAcceptRejectButtonsHTML type={toolName} codeStr={codeStr} applyBoxId={applyBoxId} uri={uri} threadId={threadId} />
-	</div>
-}
-
-
-
-const InvalidTool = ({ toolName, message, mcpServerName }: { toolName: ToolName, message: string, mcpServerName: string | undefined }) => {
-	const accessor = useAccessor()
-	const title = getTitle({ name: toolName, type: 'invalid_params', mcpServerName })
-	const desc1 = 'Invalid parameters'
-	const icon = null
-	const isError = true
-	const componentParams: ToolHeaderParams = { title, desc1, isError, icon }
-
-	componentParams.children = <ToolChildrenWrapper>
-		<CodeChildren className='bg-void-bg-3'>
-			{message}
-		</CodeChildren>
-	</ToolChildrenWrapper>
-	return <ToolHeaderWrapper {...componentParams} />
-}
-
-const CanceledTool = ({ toolName, mcpServerName }: { toolName: ToolName, mcpServerName: string | undefined }) => {
-	const accessor = useAccessor()
-	const title = getTitle({ name: toolName, type: 'rejected', mcpServerName })
-	const desc1 = ''
-	const icon = null
-	const isRejected = true
-	const componentParams: ToolHeaderParams = { title, desc1, icon, isRejected }
-	return <ToolHeaderWrapper {...componentParams} />
-}
-
-
-const CommandTool = ({ toolMessage, type, threadId }: { threadId: string } & ({
-	toolMessage: Exclude<ToolMessage<'run_command'>, { type: 'invalid_params' }>
-	type: 'run_command'
-} | {
-	toolMessage: Exclude<ToolMessage<'run_persistent_command'>, { type: 'invalid_params' }>
-	type: | 'run_persistent_command'
-})) => {
-	const accessor = useAccessor()
-
-	const commandService = accessor.get('ICommandService')
-	const terminalToolsService = accessor.get('ITerminalToolService')
-	const toolsService = accessor.get('IToolsService')
-	const isError = false
-	const title = getTitle(toolMessage)
-	const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-	const icon = null
-	const streamState = useChatThreadsStreamState(threadId)
-
-	const divRef = useRef<HTMLDivElement | null>(null)
-
-	const isRejected = toolMessage.type === 'rejected'
-	const { rawParams, params } = toolMessage
-	const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-
-	const effect = async () => {
-		if (streamState?.isRunning !== 'tool') return
-		if (type !== 'run_command' || toolMessage.type !== 'running_now') return;
-
-		// wait for the interruptor so we know it's running
-
-		await streamState?.interrupt
-		const container = divRef.current;
-		if (!container) return;
-
-		const terminal = terminalToolsService.getTemporaryTerminal(toolMessage.params.terminalId);
-		if (!terminal) return;
-
-		try {
-			terminal.attachToElement(container);
-			terminal.setVisible(true)
-		} catch {
-		}
-
-		// Listen for size changes of the container and keep the terminal layout in sync.
-		const resizeObserver = new ResizeObserver((entries) => {
-			const height = entries[0].borderBoxSize[0].blockSize;
-			const width = entries[0].borderBoxSize[0].inlineSize;
-			if (typeof terminal.layout === 'function') {
-				terminal.layout({ width, height });
-			}
-		});
-
-		resizeObserver.observe(container);
-		return () => { terminal.detachFromElement(); resizeObserver?.disconnect(); }
-	}
-
-	useEffect(() => {
-		effect()
-	}, [terminalToolsService, toolMessage, toolMessage.type, type]);
-
-	if (toolMessage.type === 'success') {
-		const { result } = toolMessage
-
-		// it's unclear that this is a button and not an icon.
-		// componentParams.desc2 = <JumpToTerminalButton
-		// 	onClick={() => { terminalToolsService.openTerminal(terminalId) }}
-		// />
-
-		let msg: string
-		if (type === 'run_command') msg = toolsService.stringOfResult['run_command'](toolMessage.params, result)
-		else msg = toolsService.stringOfResult['run_persistent_command'](toolMessage.params, result)
-
-		if (type === 'run_persistent_command') {
-			componentParams.info = persistentTerminalNameOfId(toolMessage.params.persistentTerminalId)
-		}
-
-		componentParams.children = <ToolChildrenWrapper className='whitespace-pre text-nowrap overflow-auto text-sm'>
-			<div className='!select-text cursor-auto'>
-				<BlockCode initValue={`${msg.trim()}`} language='shellscript' />
-			</div>
-		</ToolChildrenWrapper>
-	}
-	else if (toolMessage.type === 'tool_error') {
-		const { result } = toolMessage
-		componentParams.bottomChildren = <BottomChildren title='Error'>
-			<CodeChildren>
-				{result}
-			</CodeChildren>
-		</BottomChildren>
-	}
-	else if (toolMessage.type === 'running_now') {
-		if (type === 'run_command')
-			componentParams.children = <div ref={divRef} className='relative h-[300px] text-sm' />
-	}
-	else if (toolMessage.type === 'rejected' || toolMessage.type === 'tool_request') {
-	}
-
-	return <>
-		<ToolHeaderWrapper {...componentParams} isOpen={type === 'run_command' && toolMessage.type === 'running_now' ? true : undefined} />
-	</>
-}
-
-type WrapperProps<T extends ToolName> = { toolMessage: Exclude<ToolMessage<T>, { type: 'invalid_params' }>, messageIdx: number, threadId: string }
-const MCPToolWrapper = ({ toolMessage }: WrapperProps<string>) => {
-	const accessor = useAccessor()
-	const mcpService = accessor.get('IMCPService')
-
-	const title = getTitle(toolMessage)
-	const desc1 = removeMCPToolNamePrefix(toolMessage.name)
-	const icon = null
-
-
-	if (toolMessage.type === 'running_now') return null // do not show running
-
-	const isError = false
-	const isRejected = toolMessage.type === 'rejected'
-	const { rawParams, params } = toolMessage
-
-	// Redact sensitive values in params before display/copy
-	const redactParams = (value: any): any => {
-		const SENSITIVE_KEYS = new Set(['token', 'apiKey', 'apikey', 'password', 'authorization', 'auth', 'secret', 'clientSecret', 'accessToken', 'bearer'])
-		const redactValue = (v: any) => (typeof v === 'string' ? (v.length > 6 ? v.slice(0, 3) + '***' + v.slice(-2) : '***') : v)
-		if (Array.isArray(value)) return value.map(redactParams)
-		if (value && typeof value === 'object') {
-			const out: any = Array.isArray(value) ? [] : {}
-			for (const k of Object.keys(value)) {
-				if (SENSITIVE_KEYS.has(k.toLowerCase())) out[k] = redactValue(value[k])
-				else out[k] = redactParams(value[k])
-			}
-			return out
-		}
-		return value
-	}
-	const componentParams: ToolHeaderParams = { title, desc1, isError, icon, isRejected, }
-
-	const redactedParams = redactParams(params)
-	const paramsStr = JSON.stringify(redactedParams, null, 2)
-	componentParams.desc2 = <CopyButton codeStr={paramsStr} toolTipName={`Copy inputs (redacted): ${paramsStr}`} />
-
-	componentParams.info = !toolMessage.mcpServerName ? 'MCP tool not found' : undefined
-
-	// Add copy inputs button in desc2
-
-
-	if (toolMessage.type === 'success' || toolMessage.type === 'tool_request') {
-		const { result } = toolMessage
-		if (result) {
-			const resultStr = mcpService.stringifyResult(result)
-			// Check if result is text (not JSON) - text events return plain text, others return JSON
-			// Type guard: check if result has 'event' property and it's 'text'
-			const isTextResult = typeof result === 'object' && result !== null && 'event' in result && (result as any).event === 'text'
-			// If it's text, display as markdown; otherwise display as JSON code block
-			const displayContent = isTextResult ? resultStr : `\`\`\`json\n${resultStr}\n\`\`\``
-			componentParams.children = <ToolChildrenWrapper>
-				<SmallProseWrapper>
-					<ChatMarkdownRender
-						string={displayContent}
-						chatMessageLocation={undefined}
-						isApplyEnabled={false}
-						isLinkDetectionEnabled={true}
-					/>
-				</SmallProseWrapper>
-			</ToolChildrenWrapper>
-		}
-	}
-	else if (toolMessage.type === 'tool_error') {
-		const { result } = toolMessage
-		componentParams.bottomChildren = <BottomChildren title='Error'>
-			<CodeChildren>
-				{result}
-			</CodeChildren>
-		</BottomChildren>
-	}
-
-	return <ToolHeaderWrapper {...componentParams} />
-
-}
-
-type ResultWrapper<T extends ToolName> = (props: WrapperProps<T>) => React.ReactNode
-
-const builtinToolNameToComponent: { [T in BuiltinToolName]: { resultWrapper: ResultWrapper<T>, } } = {
-	'read_file': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-
-			const title = getTitle(toolMessage)
-
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor);
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			let range: [number, number] | undefined = undefined
-			if (toolMessage.params.startLine !== null || toolMessage.params.endLine !== null) {
-				const start = toolMessage.params.startLine === null ? `1` : `${toolMessage.params.startLine}`
-				const end = toolMessage.params.endLine === null ? `` : `${toolMessage.params.endLine}`
-				const addStr = `(${start}-${end})`
-				componentParams.desc1 += ` ${addStr}`
-				range = [params.startLine || 1, params.endLine || 1]
-			}
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor, range) }
-				if (result.hasNextPage && params.pageNumber === 1)  // first page
-					componentParams.desc2 = `(truncated after ${Math.round(MAX_FILE_CHARS_PAGE) / 1000}k)`
-				else if (params.pageNumber > 1) // subsequent pages
-					componentParams.desc2 = `(part ${params.pageNumber})`
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				// JumpToFileButton removed in favor of FileLinkText
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		},
-	},
-	'get_dir_tree': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			if (params.uri) {
-				const rel = getRelative(params.uri, accessor)
-				if (rel) componentParams.info = `Only search in ${rel}`
-			}
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				componentParams.children = <ToolChildrenWrapper>
-					<SmallProseWrapper>
-						<ChatMarkdownRender
-							string={`\`\`\`\n${result.str}\n\`\`\``}
-							chatMessageLocation={undefined}
-							isApplyEnabled={false}
-							isLinkDetectionEnabled={true}
-						/>
-					</SmallProseWrapper>
-				</ToolChildrenWrapper>
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-
-		}
-	},
-	'ls_dir': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-			const explorerService = accessor.get('IExplorerService')
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			if (params.uri) {
-				const rel = getRelative(params.uri, accessor)
-				if (rel) componentParams.info = `Only search in ${rel}`
-			}
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				componentParams.numResults = result.children?.length
-				componentParams.hasNextPage = result.hasNextPage
-				componentParams.children = !result.children || (result.children.length ?? 0) === 0 ? undefined
-					: <ToolChildrenWrapper>
-						{result.children.map((child, i) => (<ListableToolItem key={child.uri?.toString() ?? `child-${i}`}
-							name={`${child.name}${child.isDirectory ? '/' : ''}`}
-							className='w-full overflow-auto'
-							onClick={() => {
-								voidOpenFileFn(child.uri, accessor)
-								// commandService.executeCommand('workbench.view.explorer'); // open in explorer folders view instead
-								// explorerService.select(child.uri, true);
-							}}
-						/>))}
-						{result.hasNextPage &&
-							<ListableToolItem name={`Results truncated (${result.itemsRemaining} remaining).`} isSmall={true} className='w-full overflow-auto' />
-						}
-					</ToolChildrenWrapper>
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		}
-	},
-	'search_pathnames_only': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			if (params.includePattern) {
-				componentParams.info = `Only search in ${params.includePattern}`
-			}
-
-			if (toolMessage.type === 'success') {
-				const { result, rawParams } = toolMessage
-				componentParams.numResults = result.uris.length
-				componentParams.hasNextPage = result.hasNextPage
-				componentParams.children = result.uris.length === 0 ? undefined
-					: <ToolChildrenWrapper>
-						{result.uris.map((uri, i) => (<ListableToolItem key={uri.toString()}
-							name={getBasename(uri.fsPath)}
-							className='w-full overflow-auto'
-							onClick={() => { voidOpenFileFn(uri, accessor) }}
-						/>))}
-						{result.hasNextPage &&
-							<ListableToolItem name={'Results truncated.'} isSmall={true} className='w-full overflow-auto' />
-						}
-
-					</ToolChildrenWrapper>
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		}
-	},
-	'search_for_files': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			if (params.searchInFolder || params.isRegex) {
-				let info: string[] = []
-				if (params.searchInFolder) {
-					const rel = getRelative(params.searchInFolder, accessor)
-					if (rel) info.push(`Only search in ${rel}`)
-				}
-				if (params.isRegex) { info.push(`Uses regex search`) }
-				componentParams.info = info.join('; ')
-			}
-
-			if (toolMessage.type === 'success') {
-				const { result, rawParams } = toolMessage
-				componentParams.numResults = result.uris.length
-				componentParams.hasNextPage = result.hasNextPage
-				componentParams.children = result.uris.length === 0 ? undefined
-					: <ToolChildrenWrapper>
-						{result.uris.map((uri, i) => (<ListableToolItem key={uri.toString()}
-							name={getBasename(uri.fsPath)}
-							className='w-full overflow-auto'
-							onClick={() => { voidOpenFileFn(uri, accessor) }}
-						/>))}
-						{result.hasNextPage &&
-							<ListableToolItem name={`Results truncated.`} isSmall={true} className='w-full overflow-auto' />
-						}
-
-					</ToolChildrenWrapper>
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-			return <ToolHeaderWrapper {...componentParams} />
-		}
-	},
-
-	'search_in_file': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor();
-			const toolsService = accessor.get('IToolsService');
-			const title = getTitle(toolMessage);
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor);
-			const icon = null;
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const { rawParams, params } = toolMessage;
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected };
-
-			const infoarr: string[] = []
-			const uriStr = getRelative(params.uri, accessor)
-			if (uriStr) infoarr.push(uriStr)
-			if (params.isRegex) infoarr.push('Uses regex search')
-			componentParams.info = infoarr.join('; ')
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage; // result is array of snippets
-				componentParams.numResults = result.lines.length;
-				componentParams.children = result.lines.length === 0 ? undefined :
-					<ToolChildrenWrapper>
-						<CodeChildren className='bg-void-bg-3'>
-							<pre className='font-mono whitespace-pre'>
-								{toolsService.stringOfResult['search_in_file'](params, result)}
-							</pre>
-						</CodeChildren>
-					</ToolChildrenWrapper>
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage;
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />;
-		}
-	},
-
-	'read_lint_errors': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-
-			const title = getTitle(toolMessage)
-
-			const { uri } = toolMessage.params ?? {}
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			componentParams.info = getRelative(uri, accessor) // full path
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) }
-				if (result.lintErrors)
-					componentParams.children = <LintErrorChildren lintErrors={result.lintErrors} />
-				else
-					componentParams.children = `No lint errors found.`
-
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				// JumpToFileButton removed in favor of FileLinkText
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		},
-	},
-
-	// ---
-
-	'create_file_or_folder': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			componentParams.info = getRelative(params.uri, accessor) // full path
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) }
-			}
-			else if (toolMessage.type === 'rejected') {
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) }
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				if (params) { componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) } }
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-			else if (toolMessage.type === 'running_now') {
-				// nothing more is needed
-			}
-			else if (toolMessage.type === 'tool_request') {
-				// nothing more is needed
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		}
-	},
-	'delete_file_or_folder': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-			const isFolder = toolMessage.params?.isFolder ?? false
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			componentParams.info = getRelative(params.uri, accessor) // full path
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) }
-			}
-			else if (toolMessage.type === 'rejected') {
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) }
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				if (params) { componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) } }
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-			else if (toolMessage.type === 'running_now') {
-				const { result } = toolMessage
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) }
-			}
-			else if (toolMessage.type === 'tool_request') {
-				const { result } = toolMessage
-				componentParams.onClick = () => { voidOpenFileFn(params.uri, accessor) }
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		}
-	},
-	'rewrite_file': {
-		resultWrapper: (params) => {
-			return <EditTool {...params} content={params.toolMessage.params.newContent} />
-		}
-	},
-	'edit_file': {
-		resultWrapper: (params) => {
-			return <EditTool {...params} content={params.toolMessage.params.searchReplaceBlocks} />
-		}
-	},
-
-	// ---
-
-	'run_command': {
-		resultWrapper: (params) => {
-			return <CommandTool {...params} type='run_command' />
-		}
-	},
-
-	'run_persistent_command': {
-		resultWrapper: (params) => {
-			return <CommandTool {...params} type='run_persistent_command' />
-		}
-	},
-	'open_persistent_terminal': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const terminalToolsService = accessor.get('ITerminalToolService')
-
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const title = getTitle(toolMessage)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			const relativePath = params.cwd ? getRelative(URI.file(params.cwd), accessor) : ''
-			componentParams.info = relativePath ? `Running in ${relativePath}` : undefined
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				const { persistentTerminalId } = result
-				componentParams.desc1 = persistentTerminalNameOfId(persistentTerminalId)
-				componentParams.onClick = () => terminalToolsService.focusPersistentTerminal(persistentTerminalId)
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		},
-	},
-	'kill_persistent_terminal': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const commandService = accessor.get('ICommandService')
-			const terminalToolsService = accessor.get('ITerminalToolService')
-
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const title = getTitle(toolMessage)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') return null // do not show running
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			if (toolMessage.type === 'success') {
-				const { persistentTerminalId } = params
-				componentParams.desc1 = persistentTerminalNameOfId(persistentTerminalId)
-				componentParams.onClick = () => terminalToolsService.focusPersistentTerminal(persistentTerminalId)
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		},
-	},
-	'web_search': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const toolsService = accessor.get('IToolsService')
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') {
-				// Show loading indicator
-				const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError: false, icon, isRejected: false }
-				componentParams.children = <ToolChildrenWrapper>
-					<div className='flex items-center gap-2 text-sm text-void-fg-3'>
-						<IconLoading state="processing" inline />
-						<span>Searching the web...</span>
-					</div>
-				</ToolChildrenWrapper>
-				return <ToolHeaderWrapper {...componentParams} />
-			}
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				componentParams.numResults = result.results?.length || 0
-
-				if (result.results && result.results.length > 0) {
-					componentParams.children = <ToolChildrenWrapper>
-						<div className='space-y-3'>
-							{result.results.map((r: { title: string, snippet: string, url: string }, i: number) => (
-								<div key={r.url || i} className='border border-void-border-2 bg-void-bg-2 rounded p-3 hover:bg-void-bg-3 transition-colors'>
-									<a
-										href={r.url}
-										target='_blank'
-										rel='noopener noreferrer'
-										className='block group'
-									>
-										<div className='text-sm font-semibold text-blue-400 group-hover:text-blue-300 mb-1 line-clamp-2'>
-											{r.title}
-										</div>
-										<div className='text-xs text-void-fg-4 mb-2 truncate'>
-											{r.url}
-										</div>
-										<div className='text-sm text-void-fg-2 line-clamp-3'>
-											{r.snippet}
-										</div>
-									</a>
-								</div>
-							))}
-						</div>
-					</ToolChildrenWrapper>
-				} else {
-					componentParams.children = <ToolChildrenWrapper>
-						<div className='text-sm text-void-fg-3'>
-							No search results found.
-						</div>
-					</ToolChildrenWrapper>
-				}
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		},
-	},
-	'browse_url': {
-		resultWrapper: ({ toolMessage }) => {
-			const accessor = useAccessor()
-			const toolsService = accessor.get('IToolsService')
-			const title = getTitle(toolMessage)
-			const { desc1, desc1Info } = toolNameToDesc(toolMessage.name, toolMessage.params, accessor)
-			const icon = null
-
-			if (toolMessage.type === 'tool_request') return null // do not show past requests
-			if (toolMessage.type === 'running_now') {
-				// Show loading indicator
-				const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError: false, icon, isRejected: false }
-				componentParams.children = <ToolChildrenWrapper>
-					<div className='flex items-center gap-2 text-sm text-void-fg-3'>
-						<IconLoading state="processing" inline />
-						<span>Fetching content from URL...</span>
-					</div>
-				</ToolChildrenWrapper>
-				return <ToolHeaderWrapper {...componentParams} />
-			}
-
-			const isError = false
-			const isRejected = toolMessage.type === 'rejected'
-			const { rawParams, params } = toolMessage
-			const componentParams: ToolHeaderParams = { title, desc1, desc1Info, isError, icon, isRejected, }
-
-			if (toolMessage.type === 'success') {
-				const { result } = toolMessage
-				const urlStr = result.url || params.url
-
-				componentParams.onClick = () => {
-					if (urlStr) {
-						window.open(urlStr, '_blank', 'noopener,noreferrer')
-					}
-				}
-				componentParams.info = urlStr ? `Source: ${new URL(urlStr).hostname}` : undefined
-
-				if (result.content) {
-					const contentPreview = result.content.length > 2000
-						? result.content.substring(0, 2000) + '\n\n... (content truncated)'
-						: result.content
-
-					componentParams.children = <ToolChildrenWrapper>
-						<div className='space-y-3'>
-							{result.title && (
-								<div className='text-lg font-semibold text-void-fg-1'>
-									{result.title}
-								</div>
-							)}
-							{result.metadata?.publishedDate && (
-								<div className='text-xs text-void-fg-4'>
-									Published: {result.metadata.publishedDate}
-								</div>
-							)}
-							{urlStr && (
-								<a
-									href={urlStr}
-									target='_blank'
-									rel='noopener noreferrer'
-									className='text-sm text-blue-400 hover:text-blue-300 block truncate'
-								>
-									{urlStr}
-								</a>
-							)}
-							<div className='text-sm text-void-fg-2 whitespace-pre-wrap max-h-96 overflow-y-auto border border-void-border-2 bg-void-bg-3 rounded p-3'>
-								{contentPreview}
-							</div>
-						</div>
-					</ToolChildrenWrapper>
-				} else {
-					componentParams.children = <ToolChildrenWrapper>
-						<div className='text-sm text-void-fg-3'>
-							No content extracted from URL.
-						</div>
-					</ToolChildrenWrapper>
-				}
-			}
-			else if (toolMessage.type === 'tool_error') {
-				const { result } = toolMessage
-				componentParams.bottomChildren = <BottomChildren title='Error'>
-					<CodeChildren>
-						{result}
-					</CodeChildren>
-				</BottomChildren>
-			}
-
-			return <ToolHeaderWrapper {...componentParams} />
-		},
-	},
-};
-
-
-const Checkpoint = ({ message, threadId, messageIdx, isCheckpointGhost, threadIsRunning }: { message: CheckpointEntry, threadId: string; messageIdx: number, isCheckpointGhost: boolean, threadIsRunning: boolean }) => {
-	const accessor = useAccessor()
-	const chatThreadService = accessor.get('IChatThreadService')
-	const streamState = useFullChatThreadsStreamState()
-
-	// Subscribe to thread state changes properly
-	const chatThreadsState = useChatThreadsState()
-
-	const isRunning = useChatThreadsStreamState(threadId)?.isRunning
-	const isDisabled = useMemo(() => {
-		if (isRunning) return true
-		// Use Object.values().some() instead of Object.keys().find() for better performance
-		return Object.values(streamState).some(threadState => threadState?.isRunning)
-	}, [isRunning, streamState])
-
-	// Memoize message count lookup to avoid direct state access in render
-	const threadMessagesLength = chatThreadsState.allThreads[threadId]?.messages.length ?? 0
-
-	return <div
-		className={`flex items-center justify-center px-2 `}
-	>
-		<div
-			className={`
-                    text-xs
-                    text-void-fg-3
-                    select-none
-                    ${isCheckpointGhost ? 'opacity-50' : 'opacity-100'}
-					${isDisabled ? 'cursor-default' : 'cursor-pointer'}
-                `}
-			style={{ position: 'relative', display: 'inline-block' }} // allow absolute icon
-			onClick={() => {
-				if (threadIsRunning) return
-				if (isDisabled) return
-				chatThreadService.jumpToCheckpointBeforeMessageIdx({
-					threadId,
-					messageIdx,
-					jumpToUserModified: messageIdx === threadMessagesLength - 1
-				})
-			}}
-			{...isDisabled ? {
-				'data-tooltip-id': 'cortex-tooltip',
-				'data-tooltip-content': `Disabled ${isRunning ? 'when running' : 'because another thread is running'}`,
-				'data-tooltip-place': 'top',
-			} : {}}
-		>
-			Checkpoint
-		</div>
-	</div>
-}
-
-
-type ChatBubbleMode = 'display' | 'edit'
-type ChatBubbleProps = {
-	chatMessage: ChatMessage,
-	messageIdx: number,
-	isCommitted: boolean,
-	chatIsRunning: IsRunningType,
-	threadId: string,
-	currCheckpointIdx: number | undefined,
-	_scrollToBottom: (() => void) | null,
-}
-
-// Plan Component - Shows structured execution plan as a todo list
-const PlanComponent = React.memo(({ message, isCheckpointGhost, threadId, messageIdx }: { message: PlanMessage, isCheckpointGhost: boolean, threadId: string, messageIdx: number }) => {
-	const accessor = useAccessor()
-	const chatThreadService = accessor.get('IChatThreadService')
-	const [expandedSteps, setExpandedSteps] = useState<Set<number>>(new Set())
-	const [isCollapsed, setIsCollapsed] = useState(false)
-
-	// Subscribe to thread state changes properly
-	const chatThreadsState = useChatThreadsState()
-    const approvalState = message.approvalState || 'pending'
-    const isRunning = useChatThreadsStreamState(threadId)?.isRunning
-    const isBusy = isRunning === 'LLM' || isRunning === 'tool' || isRunning === 'preparing'
-    const isIdleLike = isRunning === undefined || isRunning === 'idle'
-
-	// Get thread messages with proper subscription
-	const thread = chatThreadsState.allThreads[threadId]
-	const threadMessages = thread?.messages ?? []
-
-	// Memoize tool message lookup map for O(1) access instead of O(n) searches
-	const toolMessagesMap = useMemo(() => {
-		const map = new Map<string, ToolMessage<any>>()
-		for (const msg of threadMessages) {
-			if (msg.role === 'tool') {
-				const toolMsg = msg as ToolMessage<any>
-				map.set(toolMsg.id, toolMsg)
-			}
-		}
-		return map
-	}, [threadMessages])
-
-	// Calculate progress - memoize to avoid recalculating on every render
-	const totalSteps = message.steps.length
-	const completedSteps = useMemo(() =>
-		message.steps.filter(s => s.status === 'succeeded' || s.status === 'skipped').length
-	, [message.steps])
-	const progressText = useMemo(() =>
-		`${completedSteps} of ${totalSteps} ${totalSteps === 1 ? 'Step' : 'Steps'} Completed`
-	, [completedSteps, totalSteps])
-
-	// Memoize hasPausedSteps to avoid recalculating on every render
-	const hasPausedSteps = useMemo(() =>
-		message.steps.some(s => s.status === 'paused')
-	, [message.steps])
-
-	const getCheckmarkIcon = (status?: StepStatus, isDisabled?: boolean) => {
-		if (isDisabled) {
-			return <div className="w-5 h-5 rounded-full border-2 border-void-fg-4 flex items-center justify-center opacity-40" />
-		}
-
-		switch (status) {
-			case 'succeeded':
-				return (
-					<div className="w-5 h-5 rounded-full border-2 border-[var(--cortex-success)] bg-[var(--cortex-success)]/20 flex items-center justify-center">
-						<Check size={12} className="text-[var(--cortex-success)]" strokeWidth={3} />
-					</div>
-				)
-			case 'failed':
-				return (
-					<div className="w-5 h-5 rounded-full border-2 border-[var(--cortex-danger)] bg-[var(--cortex-danger)]/20 flex items-center justify-center">
-						<X size={12} className="text-[var(--cortex-danger)]" strokeWidth={3} />
-					</div>
-				)
-			case 'running':
-				return (
-					<div className="w-5 h-5 rounded-full border-2 border-yellow-500 bg-yellow-500/20 flex items-center justify-center">
-						<CircleEllipsis size={12} className="text-yellow-400 animate-spin" />
-					</div>
-				)
-			case 'paused':
-				return (
-					<div className="w-5 h-5 rounded-full border-2 border-orange-500 bg-orange-500/20 flex items-center justify-center">
-						<Dot size={12} className="text-orange-400" />
-					</div>
-				)
-			case 'skipped':
-				return (
-					<div className="w-5 h-5 rounded-full border-2 border-gray-500 bg-gray-500/20 flex items-center justify-center opacity-60">
-						<Ban size={12} className="text-gray-400" />
-					</div>
-				)
-			default: // queued
-				return (
-					<div className="w-5 h-5 rounded-full border-2 border-void-fg-3 flex items-center justify-center">
-						<div className="w-1.5 h-1.5 rounded-full bg-void-fg-3 opacity-60" />
-					</div>
-				)
-		}
-	}
-
-	const toggleStepExpanded = (stepNumber: number) => {
-		setExpandedSteps(prev => {
-			const next = new Set(prev)
-			if (next.has(stepNumber)) {
-				next.delete(stepNumber)
-			} else {
-				next.add(stepNumber)
-			}
-			return next
-		})
-	}
-
-    const handleApprove = () => {
-        if (isCheckpointGhost || isBusy) return
-		chatThreadService.approvePlan({ threadId, messageIdx })
-	}
-
-	const handleReject = () => {
-        if (isCheckpointGhost || isBusy) return
-		chatThreadService.rejectPlan({ threadId, messageIdx })
-	}
-
-	const handleToggleStep = (stepNumber: number) => {
-        if (isCheckpointGhost || isBusy) return
-		chatThreadService.toggleStepDisabled({ threadId, messageIdx, stepNumber })
-	}
-
-	const getStatusBadge = (status?: StepStatus) => {
-		switch (status) {
-			case 'running':
-				return <span className="px-1.5 py-0.5 text-xs rounded bg-yellow-500/20 text-yellow-400 border border-yellow-500/30">Running</span>
-			case 'failed':
-				return <span className="px-1.5 py-0.5 text-xs rounded bg-[var(--cortex-danger)]/20 text-[var(--cortex-danger)] border border-[var(--cortex-danger)]/30">Failed</span>
-			case 'paused':
-				return <span className="px-1.5 py-0.5 text-xs rounded bg-orange-500/20 text-orange-400 border border-orange-500/30">Paused</span>
-			case 'skipped':
-				return <span className="px-1.5 py-0.5 text-xs rounded bg-gray-500/20 text-gray-400 border border-gray-500/30">Skipped</span>
-			default:
-				return null
-		}
-	}
-
-	return (
-		<div className={`${isCheckpointGhost ? 'opacity-50 pointer-events-none' : ''} my-3`}>
-			<div className="bg-void-bg-1 border border-void-border-1 rounded-lg overflow-hidden">
-				{/* Header */}
-				<div className="px-4 py-3 border-b border-void-border-1 bg-void-bg-2/30">
-					<div className="flex items-center justify-between">
-						<div className="flex items-center gap-2 flex-1 min-w-0">
-							<button
-								onClick={() => setIsCollapsed(!isCollapsed)}
-								className="flex-shrink-0 p-1 hover:bg-void-bg-2 rounded transition-colors"
-								disabled={isCheckpointGhost}
-							>
-								<ChevronRight
-									size={16}
-									className={`text-void-fg-3 transition-transform ${isCollapsed ? '' : 'rotate-90'}`}
-								/>
-							</button>
-							<div className="flex items-center gap-2 flex-1 min-w-0">
-								<h3 className="text-void-fg-1 font-medium text-sm truncate">{message.summary}</h3>
-								{approvalState === 'pending' && (
-									<span className="px-2 py-0.5 text-xs rounded bg-blue-500/20 text-blue-400 border border-blue-500/30 flex-shrink-0">
-										Pending Approval
-									</span>
-								)}
-								{approvalState === 'executing' && (
-									<span className="px-2 py-0.5 text-xs rounded bg-yellow-500/20 text-yellow-400 border border-yellow-500/30 flex items-center gap-1 flex-shrink-0">
-										<CircleEllipsis size={12} className="animate-spin" />
-										Executing
-									</span>
-								)}
-								{approvalState === 'completed' && (
-									<span className="px-2 py-0.5 text-xs rounded bg-[var(--cortex-success)]/20 text-[var(--cortex-success)] border border-[var(--cortex-success)]/30 flex items-center gap-1 flex-shrink-0">
-										<Check size={12} />
-										Completed
-									</span>
-								)}
-							</div>
-						</div>
-
-						{!isCollapsed && (
-							<div className="flex items-center gap-3 flex-shrink-0">
-								<span className="text-void-fg-3 text-xs" aria-live="polite">{progressText}</span>
-                                {approvalState === 'pending' && isIdleLike && (
-									<div className="flex gap-2">
-										<button
-											title="Reject plan"
-									aria-label="Reject plan"
-											onClick={handleReject}
-											className="px-3 py-1.5 text-xs rounded bg-[var(--cortex-danger)]/10 text-[var(--cortex-danger)] border border-[var(--cortex-danger)]/20 hover:bg-[var(--cortex-danger)]/20 transition-colors focus:outline-none focus:ring-2 focus:ring-[var(--cortex-danger)]/40"
-										>
-											Reject
-										</button>
-										<button
-											title="Approve and execute"
-									aria-label="Approve and execute plan"
-											onClick={handleApprove}
-											className="px-3 py-1.5 text-xs rounded bg-[var(--cortex-success)]/10 text-[var(--cortex-success)] border border-[var(--cortex-success)]/20 hover:bg-[var(--cortex-success)]/20 transition-colors focus:outline-none focus:ring-2 focus:ring-[var(--cortex-success)]/40"
-										>
-											Approve & Execute
-										</button>
-									</div>
-								)}
-							{approvalState === 'executing' && isBusy && (
-								<button
-									aria-label="Pause plan execution"
-										onClick={() => chatThreadService.pauseAgentExecution({ threadId })}
-										className="px-3 py-1.5 text-xs rounded bg-orange-500/10 text-orange-400 border border-orange-500/20 hover:bg-orange-500/20 transition-colors focus:outline-none focus:ring-2 focus:ring-orange-500/40"
-									>
-										Pause
-									</button>
-								)}
-							{hasPausedSteps && !isBusy && (
-								<button
-									aria-label="Resume plan execution"
-										onClick={() => chatThreadService.resumeAgentExecution({ threadId })}
-										className="px-3 py-1.5 text-xs rounded bg-[var(--cortex-success)]/10 text-[var(--cortex-success)] border border-[var(--cortex-success)]/20 hover:bg-[var(--cortex-success)]/20 transition-colors focus:outline-none focus:ring-2 focus:ring-[var(--cortex-success)]/40"
-									>
-										Resume
-									</button>
-								)}
-							</div>
-						)}
-					</div>
-				</div>
-
-				{/* Todo List */}
-				{!isCollapsed && (
-					<div className="py-2">
-						{message.steps.map((step, idx) => {
-							const isExpanded = expandedSteps.has(step.stepNumber)
-							const isDisabled = step.disabled
-							const status = step.status || 'queued'
-							const hasDetails = step.tools || step.files || step.error || step.toolCalls
-
-							return (
-								<div
-									key={step.stepNumber}
-									className={`flex items-start gap-3 px-4 py-2.5 hover:bg-void-bg-2/30 transition-colors ${
-										isDisabled ? 'opacity-50' : ''
-									} ${status === 'failed' ? 'bg-[var(--cortex-danger)]/5' : ''}`}
-								>
-									{/* Checkmark */}
-									<div className="flex-shrink-0 mt-0.5">
-										{getCheckmarkIcon(status, isDisabled)}
-									</div>
-
-									{/* Content */}
-									<div className="flex-1 min-w-0">
-										<div className="flex items-start justify-between gap-3">
-											<p className={`text-void-fg-1 text-sm flex-1 leading-relaxed ${
-												isDisabled ? 'line-through text-void-fg-3' : ''
-											} ${status === 'succeeded' ? 'text-void-fg-2' : ''}`}>
-												{step.description}
-											</p>
-
-											{/* Status Badge */}
-											{getStatusBadge(status)}
-										</div>
-
-										{/* Actions Row */}
-										{(approvalState === 'pending' || (approvalState === 'executing' && status === 'failed')) && !isCheckpointGhost && (
-											<div className="flex items-center gap-2 mt-2">
-												{approvalState === 'pending' && !isRunning && (
-										<button
-											aria-label={`${isDisabled ? 'Enable' : 'Disable'} step ${step.stepNumber}`}
-														onClick={() => handleToggleStep(step.stepNumber)}
-														className="px-2 py-0.5 text-xs rounded bg-void-bg-2 text-void-fg-2 hover:bg-void-bg-2/80 border border-void-border-1 transition-colors"
-													>
-														{isDisabled ? 'Enable' : 'Disable'}
-													</button>
-												)}
-									{approvalState === 'executing' && status === 'failed' && (
-													<>
-											<button
-												aria-label={`Retry step ${step.stepNumber}`}
-															onClick={() => chatThreadService.retryStep({ threadId, messageIdx, stepNumber: step.stepNumber })}
-															className="px-2 py-0.5 text-xs rounded bg-[var(--cortex-success)]/10 text-[var(--cortex-success)] hover:bg-[var(--cortex-success)]/20 border border-[var(--cortex-success)]/20 transition-colors"
-														>
-															Retry
-														</button>
-											<button
-												aria-label={`Skip step ${step.stepNumber}`}
-															onClick={() => chatThreadService.skipStep({ threadId, messageIdx, stepNumber: step.stepNumber })}
-															className="px-2 py-0.5 text-xs rounded bg-gray-500/10 text-gray-400 hover:bg-gray-500/20 border border-gray-500/20 transition-colors"
-														>
-															Skip
-														</button>
-														{step.checkpointIdx !== undefined && step.checkpointIdx !== null && (
-								<button
-									aria-label={`Rollback step ${step.stepNumber}`}
-									onClick={() => { if (confirm('Rollback to the checkpoint before this step?')) chatThreadService.rollbackToStep({ threadId, messageIdx, stepNumber: step.stepNumber }) }}
-																className="px-2 py-0.5 text-xs rounded bg-orange-500/10 text-orange-400 hover:bg-orange-500/20 border border-orange-500/20 transition-colors"
-															>
-																Rollback
-															</button>
-														)}
-													</>
-												)}
-											</div>
-										)}
-
-										{/* Expandable Details */}
-										{hasDetails && (
-											<button
-												onClick={() => toggleStepExpanded(step.stepNumber)}
-												className="mt-2 flex items-center gap-1 text-void-fg-3 hover:text-void-fg-2 text-xs transition-colors"
-											>
-												<ChevronRight
-													size={12}
-													className={`transition-transform ${isExpanded ? 'rotate-90' : ''}`}
-												/>
-												<span>{isExpanded ? 'Hide' : 'Show'} details</span>
-											</button>
-										)}
-
-										{/* Expanded Content */}
-										{isExpanded && hasDetails && (
-											<div className="mt-3 space-y-3 pt-3 border-t border-void-border-1">
-												{step.tools && step.tools.length > 0 && (
-													<div>
-														<div className="text-void-fg-3 text-xs mb-2 font-medium">Expected Tools:</div>
-														<div className="flex flex-wrap gap-1.5">
-															{step.tools.map((tool, i) => (
-																<span key={`${step.stepNumber}-tool-${tool}-${i}`} className="px-2 py-0.5 bg-blue-500/10 text-blue-400 text-xs rounded border border-blue-500/20">
-																	{tool}
-																</span>
-															))}
-														</div>
-													</div>
-												)}
-												{step.toolCalls && step.toolCalls.length > 0 && (
-													<div>
-											<div className="text-void-fg-3 text-xs mb-2 font-medium flex items-center gap-2">Tool Calls Executed <span className="inline-flex items-center justify-center rounded-full bg-void-bg-2 text-void-fg-3 text-[10px] px-1.5 py-0.5 border border-void-border-1">{step.toolCalls.length}</span></div>
-														<div className="space-y-1.5">
-															{step.toolCalls.map((toolId, i) => {
-																// Use memoized map for O(1) lookup instead of O(n) find
-																const toolMsg = toolMessagesMap.get(toolId)
-																if (!toolMsg) return null
-
-																const isSuccess = toolMsg.type === 'success'
-																const isError = toolMsg.type === 'tool_error'
-
-																return (
-																	<div key={toolId} className={`p-2 rounded border text-xs ${
-																		isSuccess ? 'bg-[var(--cortex-success)]/10 border-[var(--cortex-success)]/20' :
-																		isError ? 'bg-[var(--cortex-danger)]/10 border-[var(--cortex-danger)]/20' :
-																		'bg-blue-500/10 border-blue-500/20'
-																	}`}>
-																		<div className="flex items-center justify-between mb-1">
-																			<span className="font-medium text-void-fg-1">{toolMsg.name}</span>
-																			{isSuccess && <Check size={12} className="text-[var(--cortex-success)]" />}
-																			{isError && <X size={12} className="text-[var(--cortex-danger)]" />}
-																		</div>
-																		{isError && toolMsg.result && (
-																			<div className="mt-1 text-[var(--cortex-danger)] text-xs">
-																				{toolMsg.result}
-																			</div>
-																		)}
-																		{isSuccess && toolMsg.result && (
-																			<details className="mt-1">
-																				<summary className="text-void-fg-3 cursor-pointer text-xs hover:text-void-fg-2">View result</summary>
-																				<pre className="mt-1 p-2 bg-void-bg-2 rounded text-xs overflow-auto max-h-32 border border-void-border-1">
-																					{typeof toolMsg.result === 'string'
-																						? toolMsg.result
-																						: JSON.stringify(toolMsg.result, null, 2)}
-																				</pre>
-																			</details>
-																		)}
-																		{isError && toolMsg.params && (
-																			<details className="mt-1">
-																				<summary className="text-void-fg-3 cursor-pointer text-xs hover:text-void-fg-2">View params</summary>
-																				<pre className="mt-1 p-2 bg-void-bg-2 rounded text-xs overflow-auto max-h-32 border border-void-border-1">
-																					{JSON.stringify(toolMsg.params, null, 2)}
-																				</pre>
-																			</details>
-																		)}
-																	</div>
-																)
-															})}
-														</div>
-													</div>
-												)}
-												{step.files && step.files.length > 0 && (
-													<div>
-														<div className="text-void-fg-3 text-xs mb-2 font-medium">Files Affected:</div>
-														<div className="flex flex-wrap gap-1.5">
-															{step.files.map((file, i) => (
-																<span key={`${file}-${i}`} className="px-2 py-0.5 bg-purple-500/10 text-purple-400 text-xs rounded border border-purple-500/20 flex items-center gap-1">
-																	<File size={12} />
-																	{file.split('/').pop()}
-																</span>
-															))}
-														</div>
-													</div>
-												)}
-												{step.error && (
-													<div className="p-2 bg-[var(--cortex-danger)]/10 border border-[var(--cortex-danger)]/20 rounded text-[var(--cortex-danger)] text-xs flex items-start gap-2">
-														<AlertTriangle size={14} className="flex-shrink-0 mt-0.5" />
-														<span>{step.error}</span>
-													</div>
-												)}
-												{(step.startTime && step.endTime) && (
-													<div className="text-void-fg-3 text-xs">
-														Duration: {((step.endTime - step.startTime) / 1000).toFixed(1)}s
-													</div>
-												)}
-												{step.checkpointIdx !== undefined && step.checkpointIdx !== null && (
-													<div className="text-void-fg-3 text-xs">
-														Checkpoint: #{step.checkpointIdx}
-													</div>
-												)}
-											</div>
-										)}
-									</div>
-								</div>
-							)
-						})}
-					</div>
-				)}
-			</div>
-		</div>
-	);
-}, (prev, next) => {
-	// Custom comparison: only re-render if plan message, checkpoint state, or thread changes
-	return prev.message === next.message &&
-		prev.isCheckpointGhost === next.isCheckpointGhost &&
-		prev.threadId === next.threadId &&
-		prev.messageIdx === next.messageIdx
-});
-
-// Review Component - Shows summary after execution
-const ReviewComponent = ({ message, isCheckpointGhost }: { message: ReviewMessage, isCheckpointGhost: boolean }) => {
-	return (
-		<div className={`${isCheckpointGhost ? 'opacity-50' : ''} my-2`}>
-			<div className={`border rounded-lg p-4 ${
-				message.completed
-					? 'bg-[var(--cortex-success)]/10 border-[var(--cortex-success)]/30'
-					: 'bg-[var(--cortex-warning)]/10 border-[var(--cortex-warning)]/30'
-			}`}>
-				<div className="flex items-center justify-between mb-3">
-					<div className="flex items-center gap-2">
-						{message.completed ? (
-							<Check className="text-[var(--cortex-success)]" size={18} />
-						) : (
-							<AlertTriangle className="text-[var(--cortex-warning)]" size={18} />
-						)}
-						<h3 className={`font-semibold text-sm ${
-							message.completed ? 'text-green-300' : 'text-amber-300'
-						}`}>
-							{message.completed ? 'Review Complete' : 'Review: Issues Found'}
-						</h3>
-					</div>
-					{(message.executionTime || message.stepsCompleted !== undefined) && (
-						<div className="text-xs text-void-fg-3">
-							{message.executionTime && `${(message.executionTime / 1000).toFixed(1)}s`}
-							{message.stepsCompleted !== undefined && message.stepsTotal !== undefined && (
-								<span className="ml-2">
-									{message.stepsCompleted}/{message.stepsTotal} steps
-								</span>
-							)}
-						</div>
-					)}
-				</div>
-				<p className="text-void-fg-2 text-sm mb-3">{message.summary}</p>
-
-				{message.filesChanged && message.filesChanged.length > 0 && (
-					<div className="mb-3">
-						<h4 className="text-void-fg-2 text-xs font-semibold mb-2">Files Changed:</h4>
-						<div className="space-y-1">
-							{message.filesChanged.map((file, i) => (
-								<div key={file.path || i} className="flex items-center gap-2 text-xs">
-									{file.changeType === 'created' && <CirclePlus className="text-[var(--cortex-success)]" size={12} />}
-									{file.changeType === 'modified' && <Pencil className="text-blue-400" size={12} />}
-									{file.changeType === 'deleted' && <X className="text-[var(--cortex-danger)]" size={12} />}
-									<span className="text-void-fg-2">{file.path}</span>
-								</div>
-							))}
-						</div>
-					</div>
-				)}
-
-				{message.issues && message.issues.length > 0 && (
-					<div className="space-y-2 mb-3">
-						{message.issues.map((issue, i) => (
-							<div key={`${issue.severity}-${i}`} className={`flex gap-2 text-sm p-2 rounded ${
-								issue.severity === 'error' ? 'bg-[var(--cortex-danger)]/10 border border-[var(--cortex-danger)]/20' :
-								issue.severity === 'warning' ? 'bg-[var(--cortex-warning)]/10 border border-[var(--cortex-warning)]/20' :
-								'bg-blue-500/10 border border-blue-500/20'
-							}`}>
-								{issue.severity === 'error' ? (
-									<X className="text-[var(--cortex-danger)] flex-shrink-0 mt-0.5" size={16} />
-								) : issue.severity === 'warning' ? (
-									<AlertTriangle className="text-[var(--cortex-warning)] flex-shrink-0 mt-0.5" size={16} />
-								) : (
-									<Info className="text-blue-400 flex-shrink-0 mt-0.5" size={16} />
-								)}
-								<div className="flex-1">
-									<p className={`${
-										issue.severity === 'error' ? 'text-red-300' :
-										issue.severity === 'warning' ? 'text-amber-300' :
-										'text-blue-300'
-									}`}>
-										{issue.message}
-									</p>
-									{issue.file && (
-										<p className="text-void-fg-3 text-xs mt-1 flex items-center gap-1">
-											<File size={12} />
-											{issue.file}
-										</p>
-									)}
-								</div>
-							</div>
-						))}
-					</div>
-				)}
-
-				{message.nextSteps && message.nextSteps.length > 0 && (
-					<div className="mt-3 pt-3 border-t border-void-border-2">
-						<p className="text-void-fg-3 text-xs mb-2 font-medium">Recommended Next Steps:</p>
-						<ul className="space-y-1">
-							{message.nextSteps.map((step, i) => (
-								<li key={`step-${i}`} className="text-void-fg-2 text-xs flex items-start gap-2">
-									<span className="text-void-fg-4 mt-1">•</span>
-									<span>{step}</span>
-								</li>
-							))}
-						</ul>
-					</div>
-				)}
-			</div>
-		</div>
-	);
-};
-
-const ChatBubble = React.memo((props: ChatBubbleProps) => {
-	return <ErrorBoundary>
-		<div className="message-enter">
-			<_ChatBubble {...props} />
-		</div>
-	</ErrorBoundary>
-}, (prev, next) => {
-	// Custom comparison: only re-render if props actually changed
-	return prev.chatMessage === next.chatMessage &&
-		prev.messageIdx === next.messageIdx &&
-		prev.isCommitted === next.isCommitted &&
-		prev.chatIsRunning === next.chatIsRunning &&
-		prev.currCheckpointIdx === next.currCheckpointIdx &&
-		prev.threadId === next.threadId &&
-		prev._scrollToBottom === next._scrollToBottom
-})
-
-const _ChatBubble = React.memo(({ threadId, chatMessage, currCheckpointIdx, isCommitted, messageIdx, chatIsRunning, _scrollToBottom }: ChatBubbleProps) => {
-	const role = chatMessage.role
-
-	const isCheckpointGhost = messageIdx > (currCheckpointIdx ?? Infinity) && !chatIsRunning // whether to show as gray (if chat is running, for good measure just dont show any ghosts)
-
-	if (role === 'user') {
-		return <UserMessageComponent
-			chatMessage={chatMessage}
-			isCheckpointGhost={isCheckpointGhost}
-			currCheckpointIdx={currCheckpointIdx}
-			messageIdx={messageIdx}
-			_scrollToBottom={_scrollToBottom}
-		/>
-	}
-	else if (role === 'assistant') {
-		return <AssistantMessageComponent
-			chatMessage={chatMessage}
-			isCheckpointGhost={isCheckpointGhost}
-			messageIdx={messageIdx}
-			isCommitted={isCommitted}
-		/>
-	}
-	else if (role === 'tool') {
-
-		if (chatMessage.type === 'invalid_params') {
-			return <div className={`${isCheckpointGhost ? 'opacity-50' : ''}`}>
-				<InvalidTool toolName={chatMessage.name} message={chatMessage.content} mcpServerName={chatMessage.mcpServerName} />
-			</div>
-		}
-
-		const toolName = chatMessage.name
-		const isBuiltInTool = isABuiltinToolName(toolName)
-		const ToolResultWrapper = isBuiltInTool ? builtinToolNameToComponent[toolName]?.resultWrapper as ResultWrapper<ToolName>
-			: MCPToolWrapper as ResultWrapper<ToolName>
-
-		if (ToolResultWrapper)
-			return <>
-				<div className={`${isCheckpointGhost ? 'opacity-50' : ''}`}>
-					<ToolResultWrapper
-						toolMessage={chatMessage}
-						messageIdx={messageIdx}
-						threadId={threadId}
-					/>
-				</div>
-				{chatMessage.type === 'tool_request' ?
-					<div className={`${isCheckpointGhost ? 'opacity-50 pointer-events-none' : ''}`}>
-						<ToolRequestAcceptRejectButtons toolName={chatMessage.name} />
-					</div> : null}
-			</>
-		return null
-	}
-
-	else if (role === 'interrupted_streaming_tool') {
-		return <div className={`${isCheckpointGhost ? 'opacity-50' : ''}`}>
-			<CanceledTool toolName={chatMessage.name} mcpServerName={chatMessage.mcpServerName} />
-		</div>
-	}
-
-	else if (role === 'checkpoint') {
-		return <Checkpoint
-			threadId={threadId}
-			message={chatMessage}
-			messageIdx={messageIdx}
-			isCheckpointGhost={isCheckpointGhost}
-			threadIsRunning={!!chatIsRunning}
-		/>
-	}
-
-	else if (role === 'plan') {
-		return <PlanComponent
-			message={chatMessage}
-			isCheckpointGhost={isCheckpointGhost}
-			threadId={threadId}
-			messageIdx={messageIdx}
-		/>
-	}
-
-	else if (role === 'review') {
-		return <ReviewComponent
-			message={chatMessage}
-			isCheckpointGhost={isCheckpointGhost}
-		/>
-	}
-
-}, (prev, next) => {
-	// Custom comparison for _ChatBubble
-	return prev.chatMessage === next.chatMessage &&
-		prev.messageIdx === next.messageIdx &&
-		prev.isCommitted === next.isCommitted &&
-		prev.chatIsRunning === next.chatIsRunning &&
-		prev.currCheckpointIdx === next.currCheckpointIdx &&
-		prev.threadId === next.threadId &&
-		prev._scrollToBottom === next._scrollToBottom
-})
 
 const CommandBarInChat = () => {
 	const { stateOfURI: commandBarStateOfURI, sortedURIs: sortedCommandBarURIs } = useCommandBarState()
@@ -2945,43 +327,6 @@ const CommandBarInChat = () => {
 	)
 }
 
-
-
-const EditToolSoFar = ({ toolCallSoFar, }: { toolCallSoFar: RawToolCallObj }) => {
-
-	if (!isABuiltinToolName(toolCallSoFar.name)) return null
-
-	const accessor = useAccessor()
-
-	const uri = toolCallSoFar.rawParams.uri ? URI.file(toolCallSoFar.rawParams.uri) : undefined
-
-	const title = titleOfBuiltinToolName[toolCallSoFar.name].proposed
-
-	const uriDone = toolCallSoFar.doneParams.includes('uri')
-	const desc1 = <span className='flex items-center gap-1'>
-		{uriDone ?
-			getBasename(toolCallSoFar.rawParams['uri'] ?? 'unknown')
-			: `Generating`}
-		<IconLoading state="processing" inline />
-	</span>
-
-	const desc1OnClick = () => { uri && voidOpenFileFn(uri, accessor) }
-
-	// If URI has not been specified
-	return <ToolHeaderWrapper
-		title={title}
-		desc1={desc1}
-		desc1OnClick={desc1OnClick}
-	>
-		<EditToolChildren
-			uri={uri}
-			code={toolCallSoFar.rawParams.search_replace_blocks ?? toolCallSoFar.rawParams.new_content ?? ''}
-			type={'rewrite'} // as it streams, show in rewrite format, don't make a diff editor
-		/>
-		<IconLoading state="processing" inline />
-	</ToolHeaderWrapper>
-
-}
 
 
 export const SidebarChat = () => {
