+++ src/vs/workbench/contrib/cortexide/browser/react/src/sidebar-tsx/tools/ToolHeader.tsx
@@ -1,0 +1,191 @@
+/*--------------------------------------------------------------------------------------
+ *  Copyright 2025 Glass Devtools, Inc. All rights reserved.
+ *  Licensed under the Apache License, Version 2.0. See LICENSE.txt for more information.
+ *--------------------------------------------------------------------------------------*/
+
+import React, { useState } from 'react';
+import { ChevronRight, AlertTriangle, Ban, CircleEllipsis } from 'lucide-react';
+
+export type ToolHeaderParams = {
+	icon?: React.ReactNode;
+	title: React.ReactNode;
+	desc1: React.ReactNode;
+	desc1OnClick?: () => void;
+	desc2?: React.ReactNode;
+	isError?: boolean;
+	info?: string;
+	desc1Info?: string;
+	isRejected?: boolean;
+	numResults?: number;
+	hasNextPage?: boolean;
+	children?: React.ReactNode;
+	bottomChildren?: React.ReactNode;
+	onClick?: () => void;
+	desc2OnClick?: () => void;
+	isOpen?: boolean;
+	className?: string;
+}
+
+export const ToolHeaderWrapper = ({
+	icon,
+	title,
+	desc1,
+	desc1OnClick,
+	desc1Info,
+	desc2,
+	numResults,
+	hasNextPage,
+	children,
+	info,
+	bottomChildren,
+	isError,
+	onClick,
+	desc2OnClick,
+	isOpen,
+	isRejected,
+	className, // applies to the main content
+}: ToolHeaderParams) => {
+
+	const [isOpen_, setIsOpen] = useState(false);
+	const isExpanded = isOpen !== undefined ? isOpen : isOpen_
+
+	const isDropdown = children !== undefined // null ALLOWS dropdown
+	const isClickable = !!(isDropdown || onClick)
+
+	const isDesc1Clickable = !!desc1OnClick
+
+	const desc1HTML = <span
+		className={`text-void-fg-4 text-xs italic truncate ml-2
+			${isDesc1Clickable ? 'cursor-pointer hover:brightness-125 transition-all duration-150' : ''}
+		`}
+		onClick={desc1OnClick}
+		{...desc1Info ? {
+			'data-tooltip-id': 'cortex-tooltip',
+			'data-tooltip-content': desc1Info,
+			'data-tooltip-place': 'top',
+			'data-tooltip-delay-show': 1000,
+		} : {}}
+	>{desc1}</span>
+
+	return (<div className=''>
+		<div className={`w-full border border-void-border-3 rounded px-2 py-1 bg-void-bg-3 overflow-hidden ${className}`}>
+			{/* header */}
+			<div className={`select-none flex items-center min-h-[24px]`}>
+				<div className={`flex items-center w-full gap-x-2 overflow-hidden justify-between ${isRejected ? 'line-through' : ''}`}>
+					{/* left */}
+					<div // container for if desc1 is clickable
+						className='ml-1 flex items-center overflow-hidden'
+					>
+						{/* title eg "> Edited File" */}
+						<div className={`
+							flex items-center min-w-0 overflow-hidden grow
+							${isClickable ? 'cursor-pointer hover:brightness-125 transition-all duration-150' : ''}
+						`}
+							onClick={() => {
+								if (isDropdown) { setIsOpen(v => !v); }
+								if (onClick) { onClick(); }
+							}}
+						>
+							{isDropdown && (<ChevronRight
+								className={`
+								text-void-fg-3 mr-0.5 h-4 w-4 flex-shrink-0 transition-transform duration-100 ease-[cubic-bezier(0.4,0,0.2,1)]
+								${isExpanded ? 'rotate-90' : ''}
+							`}
+							/>)}
+							<span className="text-void-fg-3 flex-shrink-0">{title}</span>
+
+							{!isDesc1Clickable && desc1HTML}
+						</div>
+						{isDesc1Clickable && desc1HTML}
+					</div>
+
+					{/* right */}
+					<div className="flex items-center gap-x-2 flex-shrink-0">
+
+						{info && <CircleEllipsis
+							className='ml-2 text-void-fg-4 opacity-60 flex-shrink-0'
+							size={14}
+							data-tooltip-id='cortex-tooltip'
+							data-tooltip-content={info}
+							data-tooltip-place='top-end'
+						/>}
+
+						{isError && <AlertTriangle
+							className='text-void-warning opacity-90 flex-shrink-0'
+							size={14}
+							data-tooltip-id='cortex-tooltip'
+							data-tooltip-content={'Error running tool'}
+							data-tooltip-place='top'
+						/>}
+						{isRejected && <Ban
+							className='text-void-fg-4 opacity-90 flex-shrink-0'
+							size={14}
+							data-tooltip-id='cortex-tooltip'
+							data-tooltip-content={'Canceled'}
+							data-tooltip-place='top'
+						/>}
+						{desc2 && <span className="text-void-fg-4 text-xs" onClick={desc2OnClick}>
+							{desc2}
+						</span>}
+						{numResults !== undefined && (
+							<span className="text-void-fg-4 text-xs ml-auto mr-1">
+								{`${numResults}${hasNextPage ? '+' : ''} result${numResults !== 1 ? 's' : ''}`}
+							</span>
+						)}
+					</div>
+				</div>
+			</div>
+			{/* children */}
+			{<div
+				className={`overflow-hidden transition-all duration-200 ease-in-out ${isExpanded ? 'opacity-100 py-1' : 'max-h-0 opacity-0'}
+					text-void-fg-4 rounded-sm overflow-x-auto
+				  `}
+			//    bg-black bg-opacity-10 border border-void-border-4 border-opacity-50
+			>
+				{children}
+			</div>}
+		</div>
+		{bottomChildren}
+	</div>);
+};
+
+
+
+export const SimplifiedToolHeader = ({
+	title,
+	children,
+}: {
+	title: string;
+	children?: React.ReactNode;
+}) => {
+	const [isOpen, setIsOpen] = useState(false);
+	const isDropdown = children !== undefined;
+	return (
+		<div>
+			<div className="w-full">
+				{/* header */}
+				<div
+					className={`select-none flex items-center min-h-[24px] ${isDropdown ? 'cursor-pointer' : ''}`}
+					onClick={() => {
+						if (isDropdown) { setIsOpen(v => !v); }
+					}}
+				>
+					{isDropdown && (
+						<ChevronRight
+							className={`text-void-fg-3 mr-0.5 h-4 w-4 flex-shrink-0 transition-transform duration-100 ease-[cubic-bezier(0.4,0,0.2,1)] ${isOpen ? 'rotate-90' : ''}`}
+						/>
+					)}
+					<div className="flex items-center w-full overflow-hidden">
+						<span className="text-void-fg-3">{title}</span>
+					</div>
+				</div>
+				{/* children */}
+				{<div
+					className={`overflow-hidden transition-all duration-200 ease-in-out ${isOpen ? 'opacity-100' : 'max-h-0 opacity-0'} text-void-fg-4`}
+				>
+					{children}
+				</div>}
+			</div>
+		</div>
+	);
+};
